+++ src/main/java/model/RegStorage.java
@@ -10,7 +10,7 @@ public class RegStorage extends ByteStorage implements Registry {
   public int getRegister(String name) {
     int value = getValueAt(Registry.nameToIdx(name));
     // Sign-extend the byte value to an int
-    return (int) ((byte) value);
+    return (byte) value;
   }
 
   @Override
